feat: adds support for composing styles to styled api
