+++ packages/babel-plugin/src/css-prop/index.tsx
@@ -1,41 +1,8 @@
 import * as t from '@babel/types';
 import { NodePath } from '@babel/core';
 import { buildCompiledComponent } from '../utils/ast-builders';
-import { buildCss, CSSOutput } from '../utils/css-builders';
+import { buildCss } from '../utils/css-builders';
 import { Metadata } from '../types';
-
-const extractCssFromExpression = (node: t.Expression, meta: Metadata): CSSOutput => {
-  if (t.isIdentifier(node)) {
-    const binding = meta.parentPath.scope.getBinding(node.name);
-
-    if (
-      binding &&
-      t.isVariableDeclarator(binding.path.node) &&
-      t.isExpression(binding.path.node.init)
-    ) {
-      return buildCss(binding.path.node.init, meta);
-    }
-  }
-
-  if (t.isArrayExpression(node)) {
-    let css = '';
-    let variables: CSSOutput['variables'] = [];
-
-    node.elements.forEach((element) => {
-      if (!element) {
-        return;
-      }
-
-      const result = extractCssFromExpression(element as t.Expression, meta);
-      css += result.css;
-      variables = variables.concat(result.variables);
-    });
-
-    return { css, variables };
-  }
-
-  return buildCss(node, meta);
-};
 
 const getJsxAttributeExpression = (node: t.JSXAttribute) => {
   if (t.isStringLiteral(node.value)) {
@@ -72,9 +39,10 @@ export const visitCssPropPath = (path: NodePath<t.JSXOpeningElement>, meta: Meta
     return;
   }
 
+  const cssOutput = buildCss(getJsxAttributeExpression(cssProp), meta);
+
   // Remove css prop
   path.node.attributes.splice(cssPropIndex, 1);
-  const cssOutput = extractCssFromExpression(getJsxAttributeExpression(cssProp), meta);
 
   path.parentPath.replaceWith(
     buildCompiledComponent({
